+++ test/postgresql_extended/tables_test.cpp
@@ -506,7 +506,7 @@ ErrorCode tables_test() {
   ptree columns_id_value;
 
   constraint.put(Constraint::TYPE,
-                 static_cast<int32_t>(Constraint::ConstraintType::UNIQUE));
+                 static_cast<int64_t>(Constraint::ConstraintType::UNIQUE));
   // constraints
   constraints.push_back(std::make_pair("", constraint));
 
@@ -517,7 +517,7 @@ ErrorCode tables_test() {
   columns_id_value.clear();
   // type
   constraint.put(Constraint::TYPE,
-                 static_cast<int32_t>(Constraint::ConstraintType::CHECK));
+                 static_cast<int64_t>(Constraint::ConstraintType::CHECK));
   // columns
   columns_num_value.put("", 1);
   columns_num.push_back(std::make_pair("", columns_num_value));
